+++ src/components/MapControls.svelte
@@ -16,6 +16,7 @@
   import MapLocationControl from './MapLocationControl.svelte';
   import ViewModeControl from './ViewModeControl.svelte';
   import MapLocationDropdown from './MapLocationDropdown.svelte';
+  import Tooltip from './Tooltip.svelte';
   import {
     maps as mapsStore,
     showDisplays as showDisplaysStore,
@@ -28,8 +29,9 @@
   export let mapboxGlAccessToken;
   export let pitch;
   export let showCollisions;
-  export let viewMode;
   export let showBoundaries;
+  export let showDiff;
+  export let viewMode;
   export let zoom;
 
   const dispatch = createEventDispatcher();
@@ -42,7 +44,9 @@
   $: mapStateValidationMessages = getMapStateMessages(mapState, maps);
 
   // When showCollisions or showBoundaries changes, update map state
-  $: dispatch('mapState', { options: { showCollisions, showBoundaries } });
+  $: dispatch('mapState', {
+    options: { showCollisions, showBoundaries, showDiff },
+  });
 
   const handleGeocoderResult = ({ detail }) => {
     const { result } = detail;
@@ -97,6 +101,10 @@
   const toggleLinkLocations = () => {
     linkLocationsStore.update(value => !value);
   };
+
+  $: if (viewMode !== 'swipe' && showDiff) {
+    showDiff = false;
+  }
 </script>
 
 {#if $showDisplaysStore}
@@ -133,14 +141,22 @@
 
       <div class="control-section">
         <div class="checkboxes">
-          <label>
-            <span>Label Collisions</span>
+          <label class="checkbox-container">
+            <span class="checkbox-label">Label Collisions</span>
             <input type="checkbox" bind:checked={showCollisions} />
           </label>
-          <label>
-            <span>Tile Boundaries</span>
+          <label class="checkbox-container">
+            <span class="checkbox-label">Tile Boundaries</span>
             <input type="checkbox" bind:checked={showBoundaries} />
           </label>
+          {#if viewMode === 'swipe'}
+            <Tooltip title="Highlights visual differences between two styles.">
+              <label class="checkbox-container">
+                <span class="checkbox-label">Highlight differences</span>
+                <input type="checkbox" bind:checked={showDiff} />
+              </label>
+            </Tooltip>
+          {/if}
         </div>
       </div>
       <div class="control-section">
@@ -233,6 +249,17 @@
 
   .checkboxes {
     text-align: right;
+    font-size: 12px;
+  }
+
+  .checkbox-container {
+    display: flex;
+    align-items: center;
+    justify-content: right;
+  }
+
+  .checkbox-label {
+    margin-right: 6px;
   }
 
   .validation-messages {