+++ src/query.js
@@ -8,7 +8,7 @@ linkLocationsStore.subscribe(value => (linkedLocations = value));
 const jsonKeys = ['maps', 'locations'];
 
 // Keys that should be encoded/decoded as boolean values
-const booleanKeys = ['showCollisions', 'showBoundaries'];
+const booleanKeys = ['showCollisions', 'showBoundaries', 'showDiff'];
 
 // Keys that should be encoded (not decoded) encoded as one map param
 const mapLocationKeys = ['bearing', 'center', 'pitch', 'zoom'];